Remove scenic from recommended Gemfile entry
This gem has a dependency on `scenic` so specifying `scenic_sqlite_adapter` as a dependency will take care of this. It will also avoid any potential versioning conflicts if some future version of scenic becomes incompatible with this gem. This gem can update its `scenic` dependency in the gemspec to specify compatible versions.